added login and sign up funtion storing user data
