+++ src/App.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
-import { Authentication } from './components/firebase'
+import { Authentication } from './components/firebase';
+import { database } from './components/firebase';
 
 import { main_page } from './components/main_page';
 import { setup_page } from './components/setup_page';
@@ -26,21 +27,75 @@ export class App extends Component {
   };
 
   componentDidMount() {
-    this.authListener();
+    this.authListener();    
   }
 
+  
   authListener() {
     Authentication.auth().onAuthStateChanged((user) => {
       console.log(user);
       if (user) {
         this.setState({ user });
-        localStorage.setItem('user' , user.uid);
+        localStorage.setItem('user', user.uid);
+        this.pulldata_Health();
+        this.pulldata_Meals();
+        this.pulldata_Ingredients();
       } else {
         this.setState({ user: null })
         localStorage.removeItem('user');
       }
     });
   }
+
+  //connects to database and puts data into set variables
+  pulldata_Health() {
+      database.collection('Health_data')
+        .doc(localStorage.getItem('user'))
+        .get()
+        .then(doc => {
+          const data = doc.data();
+          localStorage.setItem('user_data', JSON.stringify(data));
+          console.log(JSON.parse(localStorage.getItem('user_data')))
+        }).catch(function (error) {
+          console.error("Error reading health", error);
+        });
+
+  }
+
+  pulldata_Meals() {
+    database.collection('Meals')
+      .doc()
+      .get()
+      .then(doc => {
+        const data = doc.data();
+        this.meals = data;
+        console.log(data);
+      })
+      .then(function () {
+        //console.log('pulled meal data')
+      })
+      .catch(function (error) {
+        console.error("Error reading meals", error);
+      });
+  }
+
+  pulldata_Ingredients() {
+    database.collection('Ingredients')
+      .doc()
+      .get()
+      .then(doc => {
+        const data = doc.data();
+        this.ingredients = data;
+        console.log(this.ingredients);
+      })
+      .then(function () {
+        //console.log('pulled ingredient data')
+      })
+      .catch(function (error) {
+        console.error("Error reading ingredients", error);
+      });
+  }
+
 
   drawerToggleClickHandler = () => {
     this.setState((prevState) => {
@@ -72,7 +127,7 @@ export class App extends Component {
             < Switch >
               <Route path='/setup_page' component={setup_page} exact />
               <Route path='/settings_page' component={settings_page} exact />
-              {this.state.user ? (<Route path='/' component={main_page} />) : (<Login_page />)}
+              {this.state.user ? (<Route path='/' component={main_page} exact />) : (<Login_page />)}
             </Switch>
 
           </div>
